// packages/openapi-ts-tests/main/test/__snapshots__/3.1.x/plugins/effect-schema/schema-const/effect-schema.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

import * as Schema from 'effect/Schema';

export const FooSchema = Schema.Struct({
  foo: Schema.optionalKey(Schema.Literal('foo')),
  bar: Schema.optionalKey(Schema.Literal(3.2).check(Schema.isFinite())),
  baz: Schema.optionalKey(Schema.Literal(-1).check(Schema.isInt())),
  qux: Schema.optionalKey(Schema.Literal(true)),
  quux: Schema.optionalKey(Schema.Tuple([
    Schema.Literal(1),
    Schema.Literal(2),
    Schema.Literal(3),
    Schema.Literal('foo'),
    Schema.Literal(true)
  ])),
  corge: Schema.optionalKey(Schema.Struct({
    foo: Schema.Literal(1).check(Schema.isInt()),
    bar: Schema.Literal(true),
    baz: Schema.Literal('grault')
  })),
  garply: Schema.optionalKey(Schema.Literal('10n').annotate({ format: 'int64' })),
  numberInt8: Schema.optionalKey(Schema.Literal(100).check(Schema.isFinite()).annotate({ format: 'int8' })),
  numberInt16: Schema.optionalKey(Schema.Literal(1000).check(Schema.isFinite()).annotate({ format: 'int16' })),
  numberInt32: Schema.optionalKey(Schema.Literal(100000).check(Schema.isFinite()).annotate({ format: 'int32' })),
  numberInt64: Schema.optionalKey(Schema.Literal(1000000000000).check(Schema.isFinite()).annotate({ format: 'int64' })),
  numberUint8: Schema.optionalKey(Schema.Literal(200).check(Schema.isFinite()).annotate({ format: 'uint8' })),
  numberUint16: Schema.optionalKey(Schema.Literal(50000).check(Schema.isFinite()).annotate({ format: 'uint16' })),
  numberUint32: Schema.optionalKey(Schema.Literal(3000000000).check(Schema.isFinite()).annotate({ format: 'uint32' })),
  numberUint64: Schema.optionalKey(Schema.Literal(18000000000000000000).check(Schema.isFinite()).annotate({ format: 'uint64' })),
  integerInt8: Schema.optionalKey(Schema.Literal(-100).check(Schema.isInt()).annotate({ format: 'int8' })),
  integerInt16: Schema.optionalKey(Schema.Literal(-1000).check(Schema.isInt()).annotate({ format: 'int16' })),
  integerInt32: Schema.optionalKey(Schema.Literal(-100000).check(Schema.isInt()).annotate({ format: 'int32' })),
  integerInt64: Schema.optionalKey(Schema.Literal(-1000000000000).check(Schema.isInt()).annotate({ format: 'int64' })),
  integerUint8: Schema.optionalKey(Schema.Literal(255).check(Schema.isInt()).annotate({ format: 'uint8' })),
  integerUint16: Schema.optionalKey(Schema.Literal(65535).check(Schema.isInt()).annotate({ format: 'uint16' })),
  integerUint32: Schema.optionalKey(Schema.Literal(4294967295).check(Schema.isInt()).annotate({ format: 'uint32' })),
  integerUint64: Schema.optionalKey(Schema.Literal('18446744073709551615n').annotate({ format: 'uint64' })),
  stringInt64: Schema.optionalKey(Schema.Literal('-9223372036854775808').annotate({ format: 'int64' })),
  stringUint64: Schema.optionalKey(Schema.Literal('18446744073709551615').annotate({ format: 'uint64' }))
});